use sharedValue directly
